+++ packaging/msi/SumoLogic.wixext/SumoLogic.wixext/CustomAction.cs
@@ -1,6 +1,7 @@
 using System;
 using System.IO;
 using System.Text.RegularExpressions;
+using System.Windows.Forms;
 using WixToolset.Dtf.WindowsInstaller;
 
 namespace SumoLogic.wixext
@@ -13,9 +14,19 @@ namespace SumoLogic.wixext
 
         // WiX property names
         private const string pCommonConfigPath = "CommonConfigPath";
+        private const string pSumoLogicConfigPath = "SumoLogicConfigPath";
         private const string pInstallationToken = "InstallationToken";
         private const string pInstallToken = "InstallToken";
         private const string pTags = "Tags";
+        private const string pApi = "Api";
+        private const string pRemotelyManaged = "RemotelyManaged";
+        private const string pEphemeral = "Ephemeral";
+        private const string pConfigFolder = "ConfigFolder";
+        private const string pOpAmpFolder = "OpAmpFolder";
+        private const string pConfigFragmentsFolder = "ConfigFragmentsFolder";
+
+        // WiX features
+        private const string fRemotelyManaged = "REMOTELYMANAGED";
 
         [CustomAction]
         public static ActionResult UpdateConfig(Session session)
@@ -26,44 +37,60 @@ namespace SumoLogic.wixext
             // Validate presence of required WiX properties
             if (!session.CustomActionData.ContainsKey(pCommonConfigPath))
             {
-                showErrorMessage(session, ecMissingCustomActionData, pCommonConfigPath);
+                ShowErrorMessage(session, ecMissingCustomActionData, pCommonConfigPath);
                 return ActionResult.Failure;
             }
             if (!session.CustomActionData.ContainsKey(pInstallToken) && !session.CustomActionData.ContainsKey(pInstallationToken))
             {
-                showErrorMessage(session, ecMissingCustomActionData, pInstallationToken);
+                ShowErrorMessage(session, ecMissingCustomActionData, pInstallationToken);
                 return ActionResult.Failure;
             }
             if (!session.CustomActionData.ContainsKey(pTags))
             {
-                showErrorMessage(session, ecMissingCustomActionData, pTags);
+                ShowErrorMessage(session, ecMissingCustomActionData, pTags);
                 return ActionResult.Failure;
             }
 
             var commonConfigPath = session.CustomActionData[pCommonConfigPath];
+            var sumoLogicConfigPath = session.CustomActionData[pSumoLogicConfigPath];
             var tags = session.CustomActionData[pTags];
 
             var installationToken = "";
-            if (session.CustomActionData.ContainsKey(pInstallationToken) && session.CustomActionData[pInstallationToken] != "") {
+            if (session.CustomActionData.ContainsKey(pInstallationToken) && session.CustomActionData[pInstallationToken] != "")
+            {
                 installationToken = session.CustomActionData[pInstallationToken];
-            } else if (session.CustomActionData.ContainsKey(pInstallToken)){
+            }
+            else if (session.CustomActionData.ContainsKey(pInstallToken))
+            {
                 installationToken = session.CustomActionData[pInstallToken];
             }
 
+            var remotelyManaged = (session.CustomActionData.ContainsKey(pRemotelyManaged) && session.CustomActionData[pRemotelyManaged] == "true");
+            var ephemeral = (session.CustomActionData.ContainsKey(pEphemeral) && session.CustomActionData[pEphemeral] == "true");
+            var opAmpFolder = session.CustomActionData[pOpAmpFolder];
+            var api = session.CustomActionData[pApi];
+
+            if (remotelyManaged && string.IsNullOrEmpty(opAmpFolder))
+            {
+                ShowErrorMessage(session, ecMissingCustomActionData, pOpAmpFolder);
+                return ActionResult.Failure;
+            }
+
             // Load config from disk and replace values
-            Config config = new Config();
-            config.InstallationToken = installationToken;
+            Config config = new Config { InstallationToken = installationToken, RemotelyManaged = remotelyManaged, Ephemeral = ephemeral,
+                OpAmpFolder = opAmpFolder, Api = api };
             config.SetCollectorFieldsFromTags(tags);
 
+            var configFile = remotelyManaged ? sumoLogicConfigPath : commonConfigPath;
             try
             {
-                ConfigUpdater configUpdater = new ConfigUpdater(new StreamReader(commonConfigPath));
+                ConfigUpdater configUpdater = new ConfigUpdater(new StreamReader(configFile));
                 configUpdater.Update(config);
-                configUpdater.Save(new StreamWriter(commonConfigPath));
+                configUpdater.Save(new StreamWriter(configFile));
             }
             catch (Exception e)
             {
-                showErrorMessage(session, ecConfigError, e.Message);
+                ShowErrorMessage(session, ecConfigError, e.Message);
                 return ActionResult.Failure;
             }
 
@@ -72,7 +99,34 @@ namespace SumoLogic.wixext
             return ActionResult.Success;
         }
 
-        private static void showErrorMessage(Session session, short errCode, string key)
+        [CustomAction]
+        public static ActionResult PopulateServiceArguments(Session session)
+        {
+            try
+            {
+                var configFolder = session.GetTargetPath(pConfigFolder);
+                var configFragmentsFolder = session.GetTargetPath(pConfigFragmentsFolder);
+                var remotelyManaged = session.Features.Contains(fRemotelyManaged) && session.Features[fRemotelyManaged].RequestState == InstallState.Local;
+
+                if (remotelyManaged)
+                {
+                    session["SERVICEARGUMENTS"] = " --remote-config \"opamp:" + configFolder + "sumologic.yaml\"";
+                }
+                else
+                {
+                    session["SERVICEARGUMENTS"] = " --config \"" + configFolder + "sumologic.yaml\" --config \"glob:" + configFragmentsFolder + "*.yaml\"";
+                }
+            }
+            catch (Exception e)
+            {
+                ShowErrorMessage(session, ecConfigError, e.Message + e.StackTrace);
+                return ActionResult.Failure;
+            }
+
+            return ActionResult.Success;
+        }
+
+        private static void ShowErrorMessage(Session session, short errCode, string key)
         {
             Record record = new Record(3);
             record.SetInteger(1, errCode);