+++ packaging/msi/SumoLogic.wixext/SumoLogicTests/ConfigTests.cs
@@ -8,7 +8,7 @@ namespace SumoLogicTests
     [TestClass]
     public class ConfigTests
     {
-        private static Random random = new Random();
+        private static readonly Random random = new Random();
 
         public static string RandomString(int length)
         {