+++ pkg/exporter/sumologicexporter/exporter_test.go
@@ -30,6 +30,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/component/componenttest"
+	"go.opentelemetry.io/collector/config/configcompression"
 	"go.opentelemetry.io/collector/config/confighttp"
 	"go.opentelemetry.io/collector/consumer/consumererror"
 	"go.opentelemetry.io/collector/exporter"
@@ -58,7 +59,7 @@ type exporterTest struct {
 
 func createTestConfig() *Config {
 	config := createDefaultConfig().(*Config)
-	config.CompressEncoding = NoCompression
+	config.HTTPClientSettings.Compression = NoCompression
 	config.LogFormat = TextFormat
 	config.MaxRequestBodySize = 20_971_520
 	config.MetricFormat = OTLPMetricFormat
@@ -292,10 +293,9 @@ func TestPartiallyFailed(t *testing.T) {
 
 func TestInvalidHTTPCLient(t *testing.T) {
 	exp, err := initExporter(&Config{
-		LogFormat:        "json",
-		MetricFormat:     "otlp",
-		CompressEncoding: "gzip",
-		TraceFormat:      "otlp",
+		LogFormat:    "json",
+		MetricFormat: "otlp",
+		TraceFormat:  "otlp",
 		HTTPClientSettings: confighttp.HTTPClientSettings{
 			Endpoint: "test_endpoint",
 			CustomRoundTripper: func(next http.RoundTripper) (http.RoundTripper, error) {
@@ -309,18 +309,6 @@ func TestInvalidHTTPCLient(t *testing.T) {
 		exp.start(context.Background(), componenttest.NewNopHost()),
 		"failed to create HTTP Client: roundTripperException",
 	)
-}
-
-func TestPushInvalidCompressor(t *testing.T) {
-	// Expect no requests
-	test := prepareExporterTest(t, createTestConfig(), nil)
-	test.exp.config.CompressEncoding = "invalid"
-
-	logs := LogRecordsToLogs(exampleLog())
-	logs.MarkReadOnly()
-
-	err := test.exp.pushLogsData(context.Background(), logs)
-	assert.EqualError(t, err, "failed to initialize compressor: invalid format: invalid")
 }
 
 func TestPushFailedBatch(t *testing.T) {
@@ -642,18 +630,6 @@ gauge_metric_name{foo="bar",remote_name="156955",url="http://another_url"} 245 1
 	}
 }
 
-func TestPushMetricsInvalidCompressor(t *testing.T) {
-	metrics := metricAndAttributesToPdataMetrics(exampleIntMetric())
-	metrics.MarkReadOnly()
-
-	// Expect no requests
-	test := prepareExporterTest(t, createTestConfig(), nil)
-	test.exp.config.CompressEncoding = "invalid"
-
-	err := test.exp.pushMetricsData(context.Background(), metrics)
-	assert.EqualError(t, err, "failed to initialize compressor: invalid format: invalid")
-}
-
 func TestMetricsPrometheusFormatMetadataFilter(t *testing.T) {
 	test := prepareExporterTest(t, createTestConfig(), []func(w http.ResponseWriter, req *http.Request){
 		func(w http.ResponseWriter, req *http.Request) {
@@ -680,10 +656,10 @@ func TestMetricsPrometheusFormatMetadataFilter(t *testing.T) {
 func Benchmark_ExporterPushLogs(b *testing.B) {
 	createConfig := func() *Config {
 		config := createDefaultConfig().(*Config)
-		config.CompressEncoding = GZIPCompression
 		config.MetricFormat = PrometheusFormat
 		config.LogFormat = TextFormat
 		config.HTTPClientSettings.Auth = nil
+		config.HTTPClientSettings.Compression = configcompression.Gzip
 		return config
 	}