+++ scripts/install.ps1
@@ -19,7 +19,16 @@ param (
     [bool] $InstallHostMetrics,
 
     # Fips is used to download a fips binary installer.
-    [bool] $Fips
+    [bool] $Fips,
+
+    # Specifies wether or not remote management is enabled
+    [bool] $RemotelyManaged,
+
+    # Ephemeral option enabled
+    [bool] $Ephemeral,
+
+    # The API URL used to communicate with the SumoLogic backend
+    [string] $Api
 )
 
 $PackageGithubOrg = "SumoLogic"
@@ -488,8 +497,17 @@ try {
         $tagsProperty = $tagStrs -Join ","
         $msiProperties += "TAGS=`"${tagsProperty}`""
     }
+    if ($Api.Length -gt 0) {
+        $msiProperties += "API=`"${Api}`""
+    }
     if ($InstallHostMetrics -eq $true) {
         $msiAddLocal += "HOSTMETRICS"
+    }
+    if ($RemotelyManaged -eq $true) {
+        $msiAddLocal += "REMOTELYMANAGED"
+    }
+    if ($Ephemeral -eq $true) {
+        $msiAddLocal += "EPHEMERAL"
     }
     if ($msiAddLocal.Count -gt 0) {
         $addLocalStr = $msiAddLocal -Join ","